refactor(project): Consolidate Django settings and quoting app views
Consolidate Django settings into a single module.
Move `PDFPriceListImportView` to `apps/quoting/views.py`.
Remove verbose debug logging from `apps/accounts` models and utils.
Update `DJANGO_SETTINGS_MODULE` path in adhoc scripts and `mypy.ini`.
Update URL documentation for the `quoting` app.

The Django settings have been unified into a single module, removing the
previous split between `base.py`, `local.py`, and `production_like.py`.
This simplifies configuration management and reduces potential for
environment-specific misconfigurations.

The `PDFPriceListImportView` and its related logic have been moved
directly into `apps/quoting/views.py`, eliminating the `views/__init__.py`
and `views/pdf_import_view.py` files. This streamlines the `quoting` app's
directory structure and simplifies module imports.

Excessive debug logging statements in `apps/accounts` have been removed
to reduce log noise and improve application performance.

Various scripts and the `mypy.ini` configuration have been updated to
reflect the new, consolidated Django settings module path, ensuring
correct environment setup and type checking.

The URL documentation has been updated to accurately reflect the changes
in view locations and descriptions.
