+++ adhoc/test_pdf_parsing.py
@@ -5,7 +5,7 @@ import sys
 import django
 
 sys.path.insert(0, os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-os.environ.setdefault("DJANGO_SETTINGS_MODULE", "jobs_manager.settings.local")
+os.environ.setdefault("DJANGO_SETTINGS_MODULE", "jobs_manager.settings")
 django.setup()
 
 from apps.quoting.services.gemini_price_list_extraction import (
