+++ apps/accounts/utils.py
@@ -1,10 +1,7 @@
-import logging
 import uuid
 from typing import Any, List, Optional
 
 from django.contrib.auth import get_user_model
-
-logger = logging.getLogger(__name__)
 
 
 def get_excluded_staff(apps_registry: Optional[Any] = None) -> List[str]:
@@ -21,41 +18,25 @@ def get_excluded_staff(apps_registry: Optional[Any] = None) -> List[str]:
         else:
             Staff = get_user_model()
 
-        logger.info(
-            "get_excluded_staff called - checking staff with invalid IMS payroll IDs"
-        )
-
         # Exclude staff members with no valid IMS payroll ID
         staff_with_ids = Staff.objects.filter(is_active=True).values_list(
             "id", "ims_payroll_id", "first_name", "last_name"
         )
 
-        logger.info(f"Found {len(staff_with_ids)} active staff members to check")
-
         for staff_id, ims_payroll_id, first_name, last_name in staff_with_ids:
             # Check for null/empty first_name
             if not first_name or first_name.strip() == "":
-                logger.warning(
-                    f"WARNING: Staff with null/empty first_name found: ID={staff_id}, first_name='{first_name}', last_name='{last_name}', ims_id={ims_payroll_id}"
-                )
+                pass  # No logging
 
             is_valid = is_valid_uuid(str(ims_payroll_id))
             if not is_valid:
                 excluded.append(str(staff_id))
-                logger.info(
-                    f"  - Excluding {first_name} {last_name} (ID: {staff_id}, IMS: {ims_payroll_id}) - invalid UUID"
-                )
             else:
-                logger.info(
-                    f"  - Including {first_name} {last_name} (ID: {staff_id}, IMS: {ims_payroll_id}) - valid UUID"
-                )
+                pass  # No logging
 
-        logger.info(
-            f"Successfully retrieved {len(excluded)} excluded staff (out of {len(staff_with_ids)} total active)."
-        )
-    except Exception as e:
-        logger.warning(f"Unable to access Staff model: {e}. No staff will be excluded.")
+    except Exception:
         # Return empty list when Staff model can't be accessed
+        pass
 
     return excluded
 
@@ -63,9 +44,7 @@ def get_excluded_staff(apps_registry: Optional[Any] = None) -> List[str]:
 def is_valid_uuid(val: Any) -> bool:
     """Check if string is a valid UUID."""
     try:
-        result = uuid.UUID(str(val))
-        logger.info(f"is_valid_uuid({val}) -> True (parsed as {result})")
+        uuid.UUID(str(val))
         return True
-    except (ValueError, AttributeError) as e:
-        logger.info(f"is_valid_uuid({val}) -> False (error: {e})")
+    except (ValueError, AttributeError):
         return False
