+++ mypy.ini
@@ -51,7 +51,7 @@ ignore_errors = true
 
 # Django plugin configuration
 [mypy.plugins.django-stubs]
-django_settings_module = jobs_manager.settings.local
+django_settings_module = jobs_manager.settings
 
 # Django models - let django-stubs infer all Field types without annotating every method
 [mypy-*.models.*]
